[Feat] #45 - 온보딩 뷰, 뷰컨 분리
