+++ DontBe-iOS/DontBe-iOS/Presentation/Onboarding/ViewControllers/OnboardingViewController.swift
@@ -18,47 +18,20 @@ final class OnboardingViewController: UIViewController {
     
     // MARK: - UI Components
     
-    private let progressImage: UIImageView = {
-        let progress = UIImageView()
-        progress.image = ImageLiterals.Onboarding.progressbar1
-        return progress
-    }()
-    
-    private let titleImage: UIImageView = {
-        let title = UIImageView()
-        title.image = ImageLiterals.Onboarding.imgOneTitle
-        title.contentMode = .scaleAspectFit
-        return title
-    }()
-    
-    private let mainImage: UIImageView = {
-        let mainImage = UIImageView()
-        mainImage.image = ImageLiterals.Onboarding.imgOne
-        mainImage.contentMode = .scaleAspectFit
-        return mainImage
-    }()
-    
-    private let backButton: UIButton = {
-        let backButton = BackButton()
-        backButton.isHidden = true
-        return backButton
-    }()
-    
-    private let nextButton: UIButton = {
-        let nextButton = CustomButton(title: StringLiterals.Button.next, backColor: .donBlack, titleColor: .donWhite)
-        return nextButton
-    }()
-    
-    private let skipButton = CustomButton(title: StringLiterals.Button.skip, backColor: .clear, titleColor: .donGray7)
+    private let originView = OnboardingView()
     
     // MARK: - Life Cycles
+    
+    override func loadView() {
+        super.loadView()
+        
+        view = originView
+    }
     
     override func viewDidLoad() {
         super.viewDidLoad()
         
         setUI()
-        setHierarchy()
-        setLayout()
         setAddTarget()
     }
     
@@ -78,96 +51,17 @@ extension OnboardingViewController {
         self.view.backgroundColor = .donGray1
     }
     
-    private func setHierarchy() {
-        self.view.addSubviews(backButton,
-                              progressImage,
-                              titleImage,
-                              mainImage,
-                              nextButton,
-                              skipButton)
-    }
-    
-    private func setLayout() {
-        backButton.snp.makeConstraints {
-            $0.top.equalToSuperview().inset(statusBarHeight + 38.adjusted)
-            $0.leading.equalToSuperview().inset(23.adjusted)
-        }
-        
-        progressImage.snp.makeConstraints {
-            $0.centerX.equalToSuperview()
-            $0.top.equalToSuperview().inset(statusBarHeight + 46.adjusted)
-            $0.width.equalTo(48.adjusted)
-            $0.height.equalTo(6.adjusted)
-        }
-        
-        if OnboardingViewController.pushCount == 0 {
-            titleImage.snp.makeConstraints {
-                $0.leading.trailing.equalToSuperview().inset(91.adjusted)
-                $0.top.equalToSuperview().inset(statusBarHeight + 90.adjustedH)
-                $0.height.equalTo(72.adjusted)
-            }
-            
-            mainImage.snp.makeConstraints {
-                $0.center.equalToSuperview()
-                $0.width.equalTo(360.adjusted)
-                $0.height.equalTo(340.adjusted)
-            }
-        } else if OnboardingViewController.pushCount == 1 {
-            titleImage.snp.makeConstraints {
-                $0.leading.trailing.equalToSuperview().inset(80.adjusted)
-                $0.top.equalToSuperview().inset(statusBarHeight + 83.adjustedH)
-                $0.height.equalTo(102.adjusted)
-            }
-            
-            mainImage.snp.makeConstraints {
-                $0.centerY.equalToSuperview().offset(38.adjusted)
-                $0.width.equalToSuperview()
-                $0.height.equalTo(230.adjusted)
-            }
-        } else {
-            titleImage.snp.makeConstraints {
-                $0.leading.trailing.equalToSuperview().inset(49.adjusted)
-                $0.top.equalToSuperview().inset(statusBarHeight + 90.adjustedH)
-                $0.height.equalTo(72.adjusted)
-            }
-            
-            mainImage.snp.makeConstraints {
-                $0.centerX.equalToSuperview()
-                $0.centerY.equalToSuperview().offset(31.adjusted)
-                $0.width.equalTo(336.adjusted)
-                $0.height.equalTo(238.adjusted)
-            }
-        }
-        
-        if loadUserData()?.isNotFirstUser == true {
-            nextButton.snp.makeConstraints {
-                $0.bottom.equalTo(self.view.safeAreaLayoutGuide).inset(91.adjusted)
-                $0.centerX.equalToSuperview()
-            }
-            
-            skipButton.snp.makeConstraints {
-                $0.top.equalTo(nextButton.snp.bottom).offset(12.adjusted)
-                $0.centerX.equalToSuperview()
-            }
-        } else {
-            nextButton.snp.makeConstraints {
-                $0.bottom.equalTo(self.view.safeAreaLayoutGuide).inset(29.adjusted)
-                $0.centerX.equalToSuperview()
-            }
-        }
-    }
-    
     private func setAddTarget() {
-        backButton.addTarget(self, action: #selector(backButtonTapped), for: .touchUpInside)
-        skipButton.addTarget(self, action: #selector(skipButtonTapped), for: .touchUpInside)
-        nextButton.addTarget(self, action: #selector(nextButtonTapped), for: .touchUpInside)
+        self.originView.backButton.addTarget(self, action: #selector(backButtonTapped), for: .touchUpInside)
+        self.originView.skipButton.addTarget(self, action: #selector(skipButtonTapped), for: .touchUpInside)
+        self.originView.nextButton.addTarget(self, action: #selector(nextButtonTapped), for: .touchUpInside)
     }
     
     private func setOnboardingView(viewController: OnboardingViewController) {
-        viewController.backButton.isHidden = false
-        viewController.progressImage.image = self.dummy[OnboardingViewController.pushCount].progress
-        viewController.titleImage.image = self.dummy[OnboardingViewController.pushCount].titleImage
-        viewController.mainImage.image = self.dummy[OnboardingViewController.pushCount].mainImage
+        viewController.originView.backButton.isHidden = false
+        viewController.originView.progressImage.image = self.dummy[OnboardingViewController.pushCount].progress
+        viewController.originView.titleImage.image = self.dummy[OnboardingViewController.pushCount].titleImage
+        viewController.originView.mainImage.image = self.dummy[OnboardingViewController.pushCount].mainImage
     }
     
     @objc
